Fix output simulated visibilities not matching WODEN.
This commit interpolates/extrapolates the flux density of each source component
for each fine channel, rather than just using the first bands' flux densities.

It also uses the right constant when updating the LST!

This commit also moves a lot of code around, and renames a couple of things. It
should be a lot cleaner than before!