Reorganise file structure.
Previous to this commit, hyperdrive code was in its own directory. Now,
it is in the main folder of the git repo, and all other workspace
members remain in their own directories. The structure introduced by
this commit is a little more natural, and more conveniently allows use
of the newly-available erfa-static feature.

This commit also makes some minor tweaks to documentation, and bumps
dependency versions. A yaml test needed to be tweaked with a newer
version of serde_yaml.