v0.2.0-alpha5
Fix precession not being applied to output uvfits. Also write a weight
of 0.0 to visibilities that are NaN.

Make the default minimum threshold for calibration convergence a little
more tolerant.

Improve the way raw MWA data is read; don't use all timesteps by default
and correctly use the number of fine channels that are present.

Display a little more info on frequency channels.