Fix shifting an RTS source list.
The shifted RTS source list was written out with sources being
alphabetically ordered, and the first source of an RTS source list is
important. When shifting an RTS source list, this commit preserves the
order of the sources within it. It does this with a dirty function that
should not be used generally.

Also remove the preservation of SHAPELET sources from RTS sources when
shifting. These sources should never be used.

Fix a broken test (GPS seconds were wrong).