Fix a bunch of issues around DI calibration sols.
The previous commits changed how flagging was handled when writing out
solutions, and I had forgotten to change the write code.

The hyperdrive format was supposed to be writing double-precision
floats, but was only writing at single precision. This allowed me to
remove the epsilon tolerance on one test elsewhere.

Tests have been written against the reading and writing of hyperdrive
and AO formats.