Provide control on plot y-range, DI progress bars
Users can now force the y-range on the amplitude gains of solutions
plots. If the min or max isn't given, it is determined from the data.

Users can now also turn off progress bars when doing DI calibration;
this helps to see error messages that would otherwise be hidden behind
them.

Re-ordering of struct members was a long time coming. rust-analyser
does this for you if the struct name is naked and not Self.