Use weights properly. Test uvfits/ms data reading.
Previously, weights from visibility data were not read consistently. Now
they are, and flags are used by making visibility weights negative. This
commit adds a measurement set and uvfits file with a flag injected for
testing.

In the process of testing that MS, uvfits and raw data return exactly
the same visibilities, lots of little bugs were exposed. Raw data
timestamps and frequencies were slightly off, raw data corrections were
applied in a different order to Birli which made the visibilities not
match the MS and uvfits (now they match exactly), and the precession
info was generated on a potentially different timestamp. MS tile
positions are slightly less accurate than metafits positions, so use
those if possible.

The weights are now also used in calibration. Add a test to ensure this
is working as expected. When using CUDA, this slows calibration down by
20-30%. For reasons I don't understand, when not using CUDA, calibration
is slowed by ~70%. Maybe the code is being optimised differently when
CUDA is on or off. In any case, it may be worth trying to improve
calibration performance.

Overhaul the inner-most code to read data from MS and uvfits; these are
hopefully more efficient now.

Also alter the calibration quality test so it's easier to adjust in the
future.