Add a benchmark on calibration.
This commit has to change a bunch of things to be `pub` and add a flag
to suppress chanblock convergence messages. A before-and-after check on
actual data suggests that these changes have not affected performance.

Also change the names of benchmarks to be in lowercase; this makes it
easier to select specific benchmarks from the command line.