Remove list->power law conversion code.
After discussion with Christene and Jack, it was decided that this
conversion code shouldn't be used, and the sky model is trusted as is.
Source lists that are in common use should be converted to the
"hyperdrive" format so that power laws are actually reported as such and
no guessing is done within the code.

Also add the ability to use N sources in the simulate-vis utility, and
modernise its help text.