Overhaul "list" flux density estimation.
The code is now simpler and allows for negative and positive
"measurements" to be included in a list (a linear fit is used rather
than a spectral index). Tracing log messages can also be emitted; these
have no cost if the log level is not that low.

This commit slightly changes expected visibility values in a test. Not
entirely sure why...