Fix a plotting bug.
Kariuki gave me a calibration solutions file that gave just the right
conditions to expose this one. The first tile in the solutions is
unflagged, but the last one is flagged. When running solutions-plot on
the file, it seemed to run forever. This seems to happen in the plotters
library when the x/y limits of a plot aren't sensible. I found that the
limits weren't sensible because a bad tile (the last one) was being
automatically selected as the reference. The solution was simply to
reverse the iterator as intended.

The limits should always be sensible now, too. Also simplify detecting
NaNs.