Move vis weighting outside calibration.
Multiply visibilities by weights before calibration and divide by
weights if writing visibilities out. This means weights are needed
during calibration and the code is able to run about 25% faster.

Note that there are relatively big (up to 1e-4) float-point errors
introduced by doing this; writing visibilities out of di-calibration
should therefore only be done as a "quick and dirty" convenience. When
solutions-apply is introduced, that will be the preferred way to obtain
calibrated visibilities.