Small fixes.
di-calibrate cannot currently write its sky model to an ms -- the help
text has been updated to account for this and a test now exists to check
that the help text is correct.

There were problems around files and directories being overwritten. I've
done the best I can but it looks like we can't ever handle detecting
whether we have permission to remove a directory ahead of time. We may
want to improve an error message associated with attempting to remove a
directory we don't have permission to, if it's not already clear enough.

A message stating where calibrated visibilities have been written to was
lost; this is now back.