Report calibration messages more coherently.
When unpacking input data, beam parameters etc. some things are useful
to report to the user. However, the order of these messages can appear
unrelated or random because of the way the code is ordered. This commit
attempts to tidy this issue by categorising message types.

Remove dipole delays having to be passed into new methods for input data
readers. This was weird to begin with and I don't understand why I
didn't think of the more obvious, simpler thing to do (report dipole
delays in the obs. context).

Add a method to the Delays enum that gives you the ideal dipole delays.

Remove time_average_factor from CalibrateParams; timeblocks already
describe this and time_average_factor was poorly named anyway.

Fix a test that needed a metafits file.