Fix a unity dipole gains bug.
Specifying --unity-dipole-gains did indeed set all the dipole gains to
1, but it didn't adjust the dipole delays. If any of the delays were 32,
hyperbeam still treated the corresponding dipoles as having a gain of 0.
This commit will use the ideal dipole delays for all tiles if unity
dipole gains are requested. Warnings are emitted if 32s remain.

Also fix a todo in the srclist-by-beam code.

Also add to a test that Dev started but couldn't complete.