Improve MS timestep metadata. Fix solutions-apply.
The original MS-timestep-reading code was buggy; if the first antenna
was flagged, the unflagged timesteps were incorrect. Improve MS reading
logic and clean up associated errors. Add new tests.

Use HashSets instead of slices for flagged tiles. In the process of
doing this, I noticed that solutions-apply was applying the wrong
solutions to some tiles when there were flagged tiles present! This
opened a rabbit hole of work...

TileBaselineMaps -> TileBaselineFlags. Still not a great name. Along
with the new name, there's a new tile_to_unflagged_auto_index_map
member. This allows for some cleaner logic. TileBaselineFlags are even
better tested than previously.

... However I still dislike this whole premise; keeping track of a
tile's "label" and its "index" is very confusing. I think it is going to
be easier to overhaul data reading methods with arguments that simply
state what tiles/channels should *not* be rejected when reading. Then,
the code only needs to check that (1) the input arrays to mutate with
the read data have an appropriate shape given the stated number of
tiles/channels and (2) that none of the tile/channel indices are bigger
than those that exist within the data. This is clear, unambiguous and
doesn't even need to acknowledge flagged tiles.

Flagged tiles are (mostly) no longer passed around on their own, the are
used to build a TileBaselineFlags, which is passed around instead. This
makes it much easier for downstream functions to get what they need
rather than having something complex propagated down many call sites.
Using TileBaselineFlags, the big bug in solutions-apply is fixed by
correctly mapping between an unflagged array index and its constituent
tile label(s).

When applying solutions, if one of the tiles of the tiles doesn't have a
solution, the output Jones matrix is now all zeros, rather than just the
input data. The weight is still forced to be negative as before.

Add a ignore-input-solutions-tile-flag argument to solutions-apply.
Include debug messages around which tiles are flagged and aren't
depending on user inputs in solutions-apply.

A new ObsContext method exists to print the statuses of tiles. As
hyperdrive now tracks which tiles are "unavailable", this should help to
keep complexity down.