Map metafits tiles to uvfits/MS tiles.
Prior to this commit, when reading from a uvfits/MS with a metafits
file, dipole information was used from the metafits but it was not
re-arranged to match the uvfits/MS tile order. Now, the map between
uvfits/MS tile order is created based on tile names; if there is any
mismatch, a warning is emitted and metafits dipole information is
"blindly" applied as before.