Tidy up the formatting of dipole-gains.
This command used to be confusing in that it used the 0-128 index of a
tile in what looked like the tile name. No longer -- the index and tile
name are now both printed.

Also allow verbosity to passed in, in case the user wants to see
something wrong with their metafits file.