Source list improvements.
Before this commit, hyperdrive json or yaml source lists were deserialised into
temporary types, which were then converted into the proper types used throughout
hyperdrive (e.g. `SourceComponent`). This commit directly deserialises these
files into the proper types. Performance is slightly better, but json and yaml
are still very slow to deserialise in general.

These source lists are also now directly serialised when writing to disk.

Support AO source lists 1.0 and 1.1. 1.0 uses instrumental pols, whereas 1.1
uses Stokes. Sigh.

RTS source lists with multiple SHAPELET components are now removed properly.

Boxed slices are used instead of Vecs for components inside a source, and for
shapelet coefficients. This saves some bytes in memory, because the capacity
part of a vector is not kept when convert to a boxed slice. It's also a way of
signalling to the rest of the code that sources and their components aren't
intended to be changed.