Add a test for multiple-timeblock calibration.
I thought that for a long time the solutions associated with
multiple-timeblock DI calibration were bugged in that an identity Jones
matrix was added to the solutions of the first pass through. The code
works as intended, but I've added a test here to help ensure that it
continues to behave.