Improvements to modelling code.
Objects implementing `SkyModeller` no longer need to be borrowed mutably to do
modelling. `SkyModellerCpu` never needed to be borrowed mutably, but
`SkyModellerCuda` did, because it owned the device buffers for UVWs, beam
responses and visibilities. The reason for moving the ownership of these buffers
was to pave the way for peeling re-using the buffers.

A bunch of things have had the `unsafe` annotation removed. This is due to the
CUDA API being good about yielding errors when they happen, rather than
segfaulting.

Clean up some shapelet code. Simulating shapelets with CUDA used to involve
copying shapelet constants to the device as if they were variable; now the
constants are defined in the device code and are checked against the host code
at compile time. Also deal with a rare index issue.

Fixed the behaviour of curved power laws; the CUDA implementation was wrong
because their spectral indices weren't being adjusted before they were copied to
the GPU.

Model test code is now *much* easier to read.

Moved shapelet things into the src/model directory.