Tweak cargo compilation profiles.
Previously, `cargo test` was pretty slow to compile because it was being
optimised at level 2. This commit optimises at level 0, so it's much faster to
compile, but the tests take a bit longer to complete. Just test what you need
to, I guess.

Also add a "production" profile, for maximum performance.

Adjust the CI scripts to test with optimisation, because the tests would
otherwise take an extremely long time.