Fix raw data timestamp bug.
For some obsids (large ones?), treating an mwalib GPS time as a float, adding
half an integration time and dividing by 1000 to get units of seconds was enough
to introduce a float precision issue. hifitime allows us to use its types from
nanoseconds, so we keep things as integers and fix these issues for good. This
commit also changes a couple of other potentially-problematic bits of code.