Allow iono subtraction in multiple passes.
This improves the quality of iono subtraction, because for each pass, the
residuals for all sources are smaller.

The main work in this commit is (1) rotating the low-res model before adding it
to the low-res residuals, if the iono consts are non-zero, and (2) using the
"old" iono consts (i.e. the consts at the start of the iono sub routine for a
particular source) when adding the high-res model to the high-res residuals,
before subtracting the now-better-updated high-res model from the high-res
residuals.