Make MS reading more robust.
Both WEIGHT_SPECTRUM and WEIGHT columns are searched for in the main table, and
whichever is available is used.

Users can now specify which data column to use when reading from a MS. This
always defaulted to DATA, but now another column can be used instead.

There are no more shenanigans in attempting to determine the MS array position
from the writer, instead, the average antenna position is used. This change
necessitated altering a sensitive test; the floating-point precision errors
significantly affect visibility and calibration solution values, unless the
array position was forced to a specific value.