Improve UX with a new expensive_op function.
As stated in the docs: Perform this expensive operation as a normal Rust
function, but if it takes more than a certain amount of time, display a message
to the user that you're still waiting for this operation to complete.

This function is more or less the same as one I wrote for raw-data reading ages
ago, but now I've made it generic to be used for other routines.