Improve plotting code.
The legend names have changed; g_x (gain) instead of XX, D_x (leakage) instead
of XY etc. is a "more correct" representation of the solutions.

The way that the rows and columns are distributed has been tweaked, but manual
control is also now available.

Plots can be written to a specific directory, not only the CWD. Fixes #18.