Support input data with the alternative baseline layout.
The MWA Rust libraries subscribe to the "antenna1 - antenna2" baseline layout,
but others like to use "antenna2 - antenna1" instead. This manifests in uvfits
and MS data as the negative UVW coordinates of what we would normally
generate. Not detecting and altering behaviour based off of this baseline layout
difference means that any simulated visibilities are reflected in the UV
plane.

When the alternative baseline order is detected, any visibilities that are read
are conjugated.

The baseline layout is detected by (1) finding the first UVW coordinates within
the data, (2) comparing it to the UVW we would generate from the provided
XYZs, (3) comparing it to the UVW we would generate from the XYZs if they were
negated, and (4) using the smallest difference in UVW to inform what the
baseline layout is.