Detect and use the CRAM tile with a custom beam response.
The way this works is to augment the existing de-duplication code by saying that
the CRAM tile is a unique tile, then writing its beam responses to the array of
all responses.

One limitation of this approach is the number of frequencies used in the
responses; the FEE beam typically only has 24 unique frequencies in an
observation, meaning the array of all beam responses only has a length of 24 for
the frequency dimension. But the CRAM uses an analytic beam, which defines a
different response for every frequency. Therefore using the FEE beam with a CRAM
tile will make the CRAM beam responses much coarser than they could be.

This work isn't thoroughly tested, as I've run out of time.