Fix #59 multi-timeblock solution slicing (#60)
* Fix #59 multi-timeblock solution slicing
* use wget for the beam to capture failure

Previously, in multi-timeblock calibration, an initial calibration round was done on all timesteps, and then the solutions for the individual timeblocks were initialized to the all timeblocks solution.

This seems like a sensible thing to do, but for some reason Mitchcal does not like this, and I can't figure out why!

Now, timeblock solutions are initialized to identity, and this works better.

webservices went down, so the CI was edited to ensure that http errors are propagated to the CI, rather than attempting to run the test suite with a corrupted beam file.